// backend/src/db/schemas/posthog-aggregated-events.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const PosthogAggregatedEventsSchema = z.object({
  id: z.string().uuid(),
  distinctId: z.string(),
  event: z.string(),
  properties: z.unknown(),
  organizationId: z.string().nullable().optional(),
  createdAt: z.date(),
  updatedAt: z.date()
});

export type TPosthogAggregatedEvents = z.infer<typeof PosthogAggregatedEventsSchema>;
export type TPosthogAggregatedEventsInsert = Omit<z.input<typeof PosthogAggregatedEventsSchema>, TImmutableDBKeys>;
export type TPosthogAggregatedEventsUpdate = Partial<
  Omit<z.input<typeof PosthogAggregatedEventsSchema>, TImmutableDBKeys>
>;